slow condition query handler registry
